src/infrastructure/adk_integration.py
"""
Google ADK集成

Agent开发框架集成
"""

import logging
from typing import Dict, List, Optional, Any
import google.generativeai as genai
from google.generativeai.types import HarmCategory, HarmBlockThreshold

logger = logging.getLogger(__name__)

class ReactAgent:
    """ReAct智能体实现"""
    
    def __init__(self, system_prompt: str, tools: List[str], model: str = "gemini-pro"):
        self.system_prompt = system_prompt
        self.tools = tools
        self.model = model
        self.conversation_history = []
        self._initialize_model()
    
    def _initialize_model(self):
        """初始化模型"""
        try:
            # 配置生成式AI
            genai.configure(api_key="your-api-key")  # 从配置中获取
            
            # 创建生成模型
            self.generation_config = {
                "temperature": 0.7,
                "top_p": 0.8,
                "top_k": 40,
                "max_output_tokens": 2048,
            }
            
            # 安全设置
            self.safety_settings = {
                HarmCategory.HARM_CATEGORY_HARASSMENT: HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE,
                HarmCategory.HARM_CATEGORY_HATE_SPEECH: HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE,
                HarmCategory.HARM_CATEGORY_SEXUALLY_EXPLICIT: HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE,
                HarmCategory.HARM_CATEGORY_DANGEROUS_CONTENT: HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE,
            }
            
            self.model_instance = genai.GenerativeModel(
                model_name=self.model,
                generation_config=self.generation_config,
                safety_settings=self.safety_settings
            )
            
            logger.info(f"Initialized ReactAgent with model {self.model}")
            
        except Exception as e:
            logger.error(f"Failed to initialize ReactAgent: {e}")
            raise
    
    async def execute(self, context: str) -> Dict:
        """执行Agent任务"""
        try:
            # 构建完整的提示
            full_prompt = self._build_full_prompt(context)
            
            # 调用模型生成响应
            response = await self._generate_response(full_prompt)
            
            # 处理响应
            result = self._process_response(response)
            
            # 记录对话历史
            self.conversation_history.append({
                "context": context,
                "response": response,
                "result": result
            })
            
            return result
            
        except Exception as e:
            logger.error(f"Error executing ReactAgent: {e}")
            return {"success": False, "error": str(e)}
    
    def _build_full_prompt(self, context: str) -> str:
        """构建完整的提示"""
        tools_description = "\n".join([f"- {tool}" for tool in self.tools])
        
        return f"""
{self.system_prompt}

可用工具：
{tools_description}

请根据以下上下文执行任务：

{context}

请按照以下格式响应：
1. 思考：分析当前情况
2. 行动：选择合适的工具和参数
3. 观察：分析行动结果
4. 结论：总结执行结果

请开始执行：
"""
    
    async def _generate_response(self, prompt: str) -> str:
        """生成响应"""
        try:
            # 使用异步方式调用模型
            response = await self.model_instance.generate_content_async(prompt)
            return response.text
        except Exception as e:
            logger.error(f"Error generating response: {e}")
            raise
    
    def _process_response(self, response: str) -> Dict:
        """处理响应"""
        # 解析响应，提取工具调用和结果
        # 简化实现，实际需要更复杂的解析逻辑
        return {
            "success": True,
            "response": response,
            "tools_used": self._extract_tools_used(response),
            "result": self._extract_result(response)
        }
    
    def _extract_tools_used(self, response: str) -> List[str]:
        """提取使用的工具"""
        # 简化实现，从响应中提取工具名称
        tools_used = []
        for tool in self.tools:
            if tool in response:
                tools_used.append(tool)
        return tools_used
    
    def _extract_result(self, response: str) -> str:
        """提取执行结果"""
        # 简化实现，返回最后一部分作为结果
        lines = response.strip().split('\n')
        if lines:
            return lines[-1]
        return "执行完成"

class ADKIntegration:
    """Google ADK集成"""
    
    def __init__(self, api_key: str):
        self.api_key = api_key
        self.agents = {}
        self._initialize_adk()
    
    def _initialize_adk(self):
        """初始化ADK"""
        try:
            genai.configure(api_key=self.api_key)
            logger.info("ADK initialized successfully")
        except Exception as e:
            logger.error(f"Failed to initialize ADK: {e}")
            raise
    
    def create_react_agent(self, config: Dict) -> ReactAgent:
        """创建ReAct Agent"""
        try:
            agent = ReactAgent(
                system_prompt=config["system_prompt"],
                tools=config["tools"],
                model=config.get("model", "gemini-pro")
            )
            
            agent_id = config.get("agent_id", f"agent_{len(self.agents)}")
            self.agents[agent_id] = agent
            
            logger.info(f"Created ReactAgent: {agent_id}")
            return agent
            
        except Exception as e:
            logger.error(f"Failed to create ReactAgent: {e}")
            raise
    
    def get_agent(self, agent_id: str) -> Optional[ReactAgent]:
        """获取Agent"""
        return self.agents.get(agent_id)
    
    def list_agents(self) -> List[str]:
        """列出所有Agent"""
        return list(self.agents.keys())
    
    async def execute_agent(self, agent_id: str, context: str) -> Dict:
        """执行Agent"""
        agent = self.get_agent(agent_id)
        if not agent:
            raise ValueError(f"Agent {agent_id} not found")
        
        return await agent.execute(context)
