requirements.txt
# Web框架
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# 缓存
redis==5.0.1

# 数据处理
pydantic==2.5.0
pydantic-settings==2.1.0

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1

# Google ADK
google-generativeai==0.3.2

# HTTP客户端
httpx==0.25.2

# 日志
structlog==23.2.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 监控
prometheus-client==0.19.0
